betterment

+++ inc/kul/except.hpp
@@ -42,16 +42,16 @@ namespace kul{
 class Exception : public std::runtime_error{
 	protected:
 		const char* f;
-		const int l;
+		const uint16_t l;
 		const std::exception_ptr ep;
 	public:
 		~Exception() KNOEXCEPT{}
-		Exception(const char*f, const int& l, const std::string& s) : std::runtime_error(s), f(f), l(l), ep(std::current_exception()){}
+		Exception(const char*f, const uint16_t& l, const std::string& s) : std::runtime_error(s), f(f), l(l), ep(std::current_exception()){}
 		Exception(const Exception& e) : std::runtime_error(e.what()), f(e.file()),  l(e.line()), ep(e.ep) {}
 
 		const std::string debug() 			const { return std::string(std::string(f) + " : " + std::to_string(l) + " : " + std::string(what()));}
 		const char* file() 					const { return f;}
-		const int& line() 					const { return l;}
+		const uint16_t& line() 					const { return l;}
 		const std::exception_ptr& cause() 	const { return ep;}
 		const std::string stack()	const {
 			std::stringstream ss;
@@ -68,11 +68,11 @@ class Exception : public std::runtime_error{
 
 class Exit : public Exception{
 	private:		
-		const int e;
+		const uint16_t e;
 	public:
-		Exit(const char*f, const int& l, const std::string& s, const int& e) : Exception(f, l, s), e(e){}
+		Exit(const char*f, const uint16_t& l, const std::string& s, const uint16_t& e) : Exception(f, l, s), e(e){}
 		Exit(const Exit& e) : Exception(*this), e(e.e){}		
-		const int& code() const { return e; }
+		const uint16_t& code() const { return e; }
 };
 
 #define KEXCEPT(c, m) throw c(__FILE__, __LINE__, m)
